deploy/requirements.txt
#Let's install all the necessary libraries, specifying their versions
Flask==2.2.2
Werkzeug==2.2.2
tensorflow==2.12.0
opencv-python==4.8.1.78
numpy==1.23.1
Pillow==9.3.0
gunicorn